requirements.txt
django>3,<4
short_url
captcha==0.3
cryptography>=2.8
djangorestframework
markdown

# openAPI schema v3
uritemplate
pyyaml

# form
django-form-builder==0.13
